+++ common/build.gradle
@@ -18,6 +18,9 @@ dependencies {
 
     compile 'com.google.code.gson:gson'
 
+    // used for json parsing
+    compile 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.8.7'
+
     testCompile 'org.springframework:spring-test'
     testCompile 'org.springframework.boot:spring-boot-starter-test'
 }
