+++ common/src/main/java/org/opentestsystem/rdw/ingest/common/util/JsonOrganizationParser.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.rdw.ingest.processor.service.impl;
+package org.opentestsystem.rdw.ingest.common.util;
 
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -8,7 +8,6 @@ import org.opentestsystem.rdw.ingest.common.model.DistrictGroup;
 import org.opentestsystem.rdw.ingest.common.model.Organization;
 import org.opentestsystem.rdw.ingest.common.model.School;
 import org.opentestsystem.rdw.ingest.common.model.SchoolGroup;
-import org.opentestsystem.rdw.ingest.common.util.ParserHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
