+++ common/src/main/java/org/opentestsystem/rdw/ingest/common/util/OrganizationParser.java
@@ -1,7 +1,6 @@
-package org.opentestsystem.rdw.ingest.processor.service.impl;
+package org.opentestsystem.rdw.ingest.common.util;
 
 import org.opentestsystem.rdw.ingest.common.model.School;
-import org.opentestsystem.rdw.ingest.common.util.ParserHelper;
 
 import java.util.Collection;
