+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/service/impl/CalpadsSchoolsParser.java
@@ -4,6 +4,7 @@ import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVRecord;
 import org.opentestsystem.rdw.ingest.common.model.District;
 import org.opentestsystem.rdw.ingest.common.model.School;
+import org.opentestsystem.rdw.ingest.common.util.OrganizationParser;
 import org.opentestsystem.rdw.ingest.common.util.ParserHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
