+++ package-processor/src/main/java/org/opentestsystem/rdw/ingest/processor/service/impl/DefaultOrganizationProcessor.java
@@ -2,6 +2,8 @@ package org.opentestsystem.rdw.ingest.processor.service.impl;
 
 import org.opentestsystem.rdw.ingest.common.model.ImportException;
 import org.opentestsystem.rdw.ingest.common.model.School;
+import org.opentestsystem.rdw.ingest.common.util.JsonOrganizationParser;
+import org.opentestsystem.rdw.ingest.common.util.OrganizationParser;
 import org.opentestsystem.rdw.ingest.processor.repository.SchoolRepository;
 import org.opentestsystem.rdw.ingest.common.util.DataElementErrorCollector;
 import org.opentestsystem.rdw.ingest.common.util.ParserHelper;
