+++ package-processor/src/test/java/org/opentestsystem/rdw/ingest/processor/service/impl/JsonOrganizationParserTest.java
@@ -6,6 +6,7 @@ import org.opentestsystem.rdw.ingest.common.model.District;
 import org.opentestsystem.rdw.ingest.common.model.School;
 import org.opentestsystem.rdw.ingest.common.model.SchoolGroup;
 import org.opentestsystem.rdw.ingest.common.util.DataElementErrorCollector;
+import org.opentestsystem.rdw.ingest.common.util.JsonOrganizationParser;
 import org.opentestsystem.rdw.ingest.common.util.ParserHelper;
 
 import java.io.IOException;
