+++ task-service/src/test/java/org/opentestsystem/rdw/ingest/tasking/repository/ArtClientAST.java
@@ -10,7 +10,7 @@ import org.opentestsystem.rdw.ingest.common.model.School;
 import org.opentestsystem.rdw.ingest.common.model.SchoolGroup;
 import org.opentestsystem.rdw.ingest.common.util.DataElementErrorCollector;
 import org.opentestsystem.rdw.ingest.common.util.ParserHelper;
-import org.opentestsystem.rdw.ingest.processor.service.impl.JsonOrganizationParser;
+import org.opentestsystem.rdw.ingest.common.util.JsonOrganizationParser;
 import org.opentestsystem.rdw.ingest.tasking.UpdateOrganizationsConfiguration;
 import org.opentestsystem.rdw.ingest.tasking.service.impl.DefaultUpdateOrganizationsService;
 import org.springframework.beans.factory.annotation.Autowired;
